added support for continuing installation after login

+++ setup-workspace.sh
@@ -1,10 +1,7 @@
 #!/usr/bin/env bash
 
-echo
 echo 'Fetching the latest version of Zen Workspace'
-echo
 echo -e "git clone ${WORKSPACE_REPO} ${WORKSPACE_ROOT_FOLDER}"
-echo
 
 cd /
 git clone -q --progress ${WORKSPACE_REPO} ${WORKSPACE_ROOT_FOLDER}
@@ -18,29 +15,21 @@ fi
 
 echo "checking out branch: ${WORKSPACE_REPO_BRANCH}"
 git checkout ${WORKSPACE_REPO_BRANCH}
-sleep 5
-exit 1
-echo "pulling latest origin branch: ${WORKSPACE_REPO_BRANCH}"
-git pull origin ${WORKSPACE_REPO_BRANCH}
-echo
 echo 'Installing workspace scripts'
 install_workspace_scripts
-echo
 echo 'completed installing workspace scripts'
-echo
 echo 'Setting up default hosts'
 
 # add windows_host with the clients connected IP address to ubuntu hosts file
-manage-hosts addhost windows.host ${HOST_IP_ADDRESS}
+manage-hosts updatehost windows.host ${HOST_IP_ADDRESS}
 
 # add ubuntu servers IP address to windows under the selected ubuntu hostname
 manage-hosts win-addhost ${HOSTNAME} ${LOCAL_IP_ADDRESS}
 manage-hosts win-addhost workspace ${LOCAL_IP_ADDRESS}
-manage-hosts win-addhost workspace.zen ${LOCAL_IP_ADDRESS}
+manage-hosts win-updatehost workspace.zen ${LOCAL_IP_ADDRESS}
 
-echo
+
 echo 'Setting up workspace links to windows volumes'
-echo
 echo " -> linking workspace to ${WHITE}~/workspace${NC}"
 ln -s ${WORKSPACE_ROOT_FOLDER} ~/workspace
 echo " -> linking workspace to ${WHITE}/var/www${NC}"
@@ -59,18 +48,43 @@ then
     mv ~/.bash_helpers ~/_bash_helpers
 fi
 
-ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_aliases ~/.bash_aliases
-ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_helpers ~/.bash_helpers
-
-if [ -n "$(grep 'unset color_prompt force_color_prompt' ~/.bashrc)" ]
+if [[ -f ~/.bash_profile ]]
 then
-    echo 'Removing unsetting of color vars in bashrc which are required in .bash_aliases.'
-    sudo sed -i".bak" "/unset color_prompt force_color_prompt/d" ~/.bashrc
+    mv ~/.bash_profile ~/_bash_profile
 fi
 
-source ~/.bashrc
+if [[ -f ~/.docker_stacks ]]
+then
+    mv ~/.docker_stacks ~/_docker_stacks
+fi
 
-echo
+if [[ -f ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_helpers ]]
+then
+    ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_helpers ~/.bash_helpers
+fi
+
+if [[ -f ${WORKSPACE_ROOT_FOLDER}/config/bash/.docker_stacks ]]
+then
+    ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.docker_stacks ~/.docker_stacks
+fi
+
+if [[ -f ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_aliases ]]
+then
+    ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_aliases ~/.bash_aliases
+fi
+
+if [[ -f ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_profile ]]
+then
+    ln -s ${WORKSPACE_ROOT_FOLDER}/config/bash/.bash_profile ~/.bash_profile
+fi
+
+#if [ -n "$(grep 'unset color_prompt force_color_prompt' ~/.bashrc)" ]
+#then
+#    echo 'Removing unsetting of color vars in bashrc which are required in .bash_aliases.'
+#    sudo sed -i".bak" "/unset color_prompt force_color_prompt/d" ~/.bashrc
+#fi
+
+source ~/.bash_profile
+
 echo -e "${GREEN}New bash prompt successfully installed"
 echo
-
